Read in bam file as a data.frame

Read in bam file as a data.frame

bam_to_df(filename = NULL, region = ".", tags = NULL)

Arguments

filename
path to bam file

region
samtools region query string (i.e. chr1:100-1000)

tags
bam tags to return, supplied with type suffix, i.e. "XS:A" for character, "AS:i" for integer, and "CB:Z" for string. Bam entries with missing tags will return empty strings.

Examples

bam_file <- system.file("extdata", "small_sorted.bam", package = "kentr")
res <- bam_to_df(bam_file)
head(res)

#>   chrom   start     end                                            name strand
#> 1  chr1 3015399 3015499 HWI-ST1133R:8:2304:12130:31201#GAACCCATTCTTTCCC      -
#> 2  chr1 3015400 3015500 HWI-ST1133R:8:1313:13728:38795#GAACCCATTCTTTCCC      +
#> 3  chr1 3019635 3019735 HWI-ST1133R:8:1209:10428:12865#GAACCCATTCTTTCCC      +
#> 4  chr1 3019635 3019735  HWI-ST1133R:8:2315:2419:54036#GAACCCATTCTTTCCC      -
#> 5  chr1 3043511 3043611  HWI-ST1133R:8:1107:9218:64408#GAACCCATTCTTTCCC      -
#> 6  chr1 3053963 3054063  HWI-ST1133R:8:2212:2473:90987#GAACCCATTCTTTCCC      +

res <- bam_to_df(bam_file,
                 region = "chr1:3e6-3.1e6",
                 tags = c("XS:A", "AS:i", "CB:Z"))
res

#>    chrom   start     end                                            name strand
#> 1   chr1 3015399 3015499 HWI-ST1133R:8:2304:12130:31201#GAACCCATTCTTTCCC      -
#> 2   chr1 3015400 3015500 HWI-ST1133R:8:1313:13728:38795#GAACCCATTCTTTCCC      +
#> 3   chr1 3019635 3019735 HWI-ST1133R:8:1209:10428:12865#GAACCCATTCTTTCCC      +
#> 4   chr1 3019635 3019735  HWI-ST1133R:8:2315:2419:54036#GAACCCATTCTTTCCC      -
#> 5   chr1 3043511 3043611  HWI-ST1133R:8:1107:9218:64408#GAACCCATTCTTTCCC      -
#> 6   chr1 3053963 3054063  HWI-ST1133R:8:2212:2473:90987#GAACCCATTCTTTCCC      +
#> 7   chr1 3054014 3054114  HWI-ST1133R:8:2212:2473:90987#GAACCCATTCTTTCCC      -
#> 8   chr1 3058369 3058469   HWI-ST1133R:8:2309:7368:5858#GAACCCATTCTTTCCC      -
#> 9   chr1 3058645 3058745 HWI-ST1133R:8:1111:14310:29996#GAACCCATTCTTTCCC      -
#> 10  chr1 3058647 3058746 HWI-ST1133R:8:2302:13219:91951#GAACCCATTCTTTCCC      +
#> 11  chr1 3059365 3059465  HWI-ST1133R:8:1212:8780:42946#GAACCCATTCTTTCCC      +
#> 12  chr1 3068842 3068939 HWI-ST1133R:8:2215:20128:28486#GAACCCATTCTTTCCC      +
#> 13  chr1 3068851 3068939 HWI-ST1133R:8:2215:20128:28486#GAACCCATTCTTTCCC      -
#> 14  chr1 3077199 3077299  HWI-ST1133R:8:1303:9194:69620#GAACCCATTCTTTCCC      -
#> 15  chr1 3090805 3090905  HWI-ST1133R:8:1216:5060:21458#GAACCCATTCTTTCCC      +
#> 16  chr1 3090822 3090922  HWI-ST1133R:8:1216:5060:21458#GAACCCATTCTTTCCC      -
#>    XS  AS CB
#> 1  45   0   
#> 2  45   0   
#> 3  45 -18   
#> 4  45 -18   
#> 5  43   0   
#> 6  45   0   
#> 7  45   0   
#> 8  43 -12   
#> 9  45  -6   
#> 10 45 -13   
#> 11 45  -6   
#> 12 43 -15   
#> 13 43 -18   
#> 14 43   0   
#> 15 45   0   
#> 16 45   0