Find positions of query seq matches in seq column

Find positions of query seq matches in seq column

find_motifs(df, query_seq)

Arguments

df
data frame containing a header and sequence column

query_seq
sequence to identify in df$seq column, can contain regex

Value

A data.frame with chrom, start, end, and header columns. Start and end indicate start point and end point of query_seq match.

Examples

fa_path <- system.file("extdata", "test.fasta", package = "kentr")
query_seq <- "AATAAA[GTC]"

seqs <- read_genome_seq(fa_path)
find_motifs(seqs, query_seq)

#>     chrom start   end        header
#> 1    chr1 11544 11551 chr1:0-100000
#> 2    chr1 33977 33984 chr1:0-100000
#> 3    chr1 34608 34615 chr1:0-100000
#> 4    chr1 35489 35496 chr1:0-100000
#> 5    chr1 35497 35504 chr1:0-100000
#> 6    chr1 37324 37331 chr1:0-100000
#> 7    chr1 39909 39916 chr1:0-100000
#> 8    chr1 40753 40760 chr1:0-100000
#> 9    chr1 42815 42822 chr1:0-100000
#> 10   chr1 43116 43123 chr1:0-100000
#> 11   chr1 43375 43382 chr1:0-100000
#> 12   chr1 43528 43535 chr1:0-100000
#> 13   chr1 43964 43971 chr1:0-100000
#> 14   chr1 44355 44362 chr1:0-100000
#> 15   chr1 44447 44454 chr1:0-100000
#> 16   chr1 44836 44843 chr1:0-100000
#> 17   chr1 44844 44851 chr1:0-100000
#> 18   chr1 44852 44859 chr1:0-100000
#> 19   chr1 44860 44867 chr1:0-100000
#> 20   chr1 44869 44876 chr1:0-100000
#> 21   chr1 45926 45933 chr1:0-100000
#> 22   chr1 50436 50443 chr1:0-100000
#> 23   chr1 50771 50778 chr1:0-100000
#> 24   chr1 52157 52164 chr1:0-100000
#> 25   chr1 52215 52222 chr1:0-100000
#> 26   chr1 52223 52230 chr1:0-100000
#> 27   chr1 53364 53371 chr1:0-100000
#> 28   chr1 53688 53695 chr1:0-100000
#> 29   chr1 53721 53728 chr1:0-100000
#> 30   chr1 53757 53764 chr1:0-100000
#> 31   chr1 53935 53942 chr1:0-100000
#> 32   chr1 54209 54216 chr1:0-100000
#> 33   chr1 55169 55176 chr1:0-100000
#> 34   chr1 56805 56812 chr1:0-100000
#> 35   chr1 61303 61310 chr1:0-100000
#> 36   chr1 62842 62849 chr1:0-100000
#> 37   chr1 66020 66027 chr1:0-100000
#> 38   chr1 66774 66781 chr1:0-100000
#> 39   chr1 68814 68821 chr1:0-100000
#> 40   chr1 70934 70941 chr1:0-100000
#> 41   chr1 74296 74303 chr1:0-100000
#> 42   chr1 74956 74963 chr1:0-100000
#> 43   chr1 75432 75439 chr1:0-100000
#> 44   chr1 75547 75554 chr1:0-100000
#> 45   chr1 75747 75754 chr1:0-100000
#> 46   chr1 76492 76499 chr1:0-100000
#> 47   chr1 80026 80033 chr1:0-100000
#> 48   chr1 80136 80143 chr1:0-100000
#> 49   chr1 80151 80158 chr1:0-100000
#> 50   chr1 80500 80507 chr1:0-100000
#> 51   chr1 81211 81218 chr1:0-100000
#> 52   chr1 81482 81489 chr1:0-100000
#> 53   chr1 82852 82859 chr1:0-100000
#> 54   chr1 82911 82918 chr1:0-100000
#> 55   chr1 82945 82952 chr1:0-100000
#> 56   chr1 83015 83022 chr1:0-100000
#> 57   chr1 83058 83065 chr1:0-100000
#> 58   chr1 85221 85228 chr1:0-100000
#> 59   chr1 85520 85527 chr1:0-100000
#> 60   chr1 91688 91695 chr1:0-100000
#> 61   chr1 95215 95222 chr1:0-100000
#> 62   chr1 95653 95660 chr1:0-100000
#> 63   chr1 96202 96209 chr1:0-100000
#> 64   chr1 96634 96641 chr1:0-100000
#> 65   chr1 97145 97152 chr1:0-100000
#> 66   chr1 98102 98109 chr1:0-100000
#> 67   chr2 13738 13745 chr2:0-100000
#> 68   chr2 14583 14590 chr2:0-100000
#> 69   chr2 17009 17016 chr2:0-100000
#> 70   chr2 17049 17056 chr2:0-100000
#> 71   chr2 19367 19374 chr2:0-100000
#> 72   chr2 19623 19630 chr2:0-100000
#> 73   chr2 20912 20919 chr2:0-100000
#> 74   chr2 21638 21645 chr2:0-100000
#> 75   chr2 23493 23500 chr2:0-100000
#> 76   chr2 25896 25903 chr2:0-100000
#> 77   chr2 26915 26922 chr2:0-100000
#> 78   chr2 27848 27855 chr2:0-100000
#> 79   chr2 32115 32122 chr2:0-100000
#> 80   chr2 36832 36839 chr2:0-100000
#> 81   chr2 37874 37881 chr2:0-100000
#> 82   chr2 38815 38822 chr2:0-100000
#> 83   chr2 39006 39013 chr2:0-100000
#> 84   chr2 39727 39734 chr2:0-100000
#> 85   chr2 40232 40239 chr2:0-100000
#> 86   chr2 44749 44756 chr2:0-100000
#> 87   chr2 44839 44846 chr2:0-100000
#> 88   chr2 47689 47696 chr2:0-100000
#> 89   chr2 47770 47777 chr2:0-100000
#> 90   chr2 48974 48981 chr2:0-100000
#> 91   chr2 49019 49026 chr2:0-100000
#> 92   chr2 49269 49276 chr2:0-100000
#> 93   chr2 50010 50017 chr2:0-100000
#> 94   chr2 52695 52702 chr2:0-100000
#> 95   chr2 53357 53364 chr2:0-100000
#> 96   chr2 53787 53794 chr2:0-100000
#> 97   chr2 55637 55644 chr2:0-100000
#> 98   chr2 56451 56458 chr2:0-100000
#> 99   chr2 58350 58357 chr2:0-100000
#> 100  chr2 60610 60617 chr2:0-100000
#> 101  chr2 61069 61076 chr2:0-100000
#> 102  chr2 61436 61443 chr2:0-100000
#> 103  chr2 61872 61879 chr2:0-100000
#> 104  chr2 62289 62296 chr2:0-100000
#> 105  chr2 62297 62304 chr2:0-100000
#> 106  chr2 62305 62312 chr2:0-100000
#> 107  chr2 62313 62320 chr2:0-100000
#> 108  chr2 62321 62328 chr2:0-100000
#> 109  chr2 62329 62336 chr2:0-100000
#> 110  chr2 62677 62684 chr2:0-100000
#> 111  chr2 62685 62692 chr2:0-100000
#> 112  chr2 63268 63275 chr2:0-100000
#> 113  chr2 65825 65832 chr2:0-100000
#> 114  chr2 69084 69091 chr2:0-100000
#> 115  chr2 69955 69962 chr2:0-100000
#> 116  chr2 70084 70091 chr2:0-100000
#> 117  chr2 70289 70296 chr2:0-100000
#> 118  chr2 73216 73223 chr2:0-100000
#> 119  chr2 74593 74600 chr2:0-100000
#> 120  chr2 75050 75057 chr2:0-100000
#> 121  chr2 78265 78272 chr2:0-100000
#> 122  chr2 84975 84982 chr2:0-100000
#> 123  chr2 87336 87343 chr2:0-100000
#> 124  chr2 87425 87432 chr2:0-100000
#> 125  chr2 89456 89463 chr2:0-100000
#> 126  chr2 89866 89873 chr2:0-100000
#> 127  chr2 90090 90097 chr2:0-100000
#> 128  chr2 90183 90190 chr2:0-100000
#> 129  chr2 91776 91783 chr2:0-100000
#> 130  chr2 91862 91869 chr2:0-100000
#> 131  chr2 94010 94017 chr2:0-100000
#> 132  chr2 94330 94337 chr2:0-100000
#> 133  chr2 96166 96173 chr2:0-100000
#> 134  chr2 96499 96506 chr2:0-100000
#> 135  chr2 96602 96609 chr2:0-100000
#> 136  chr2 96620 96627 chr2:0-100000
#> 137  chr2 97096 97103 chr2:0-100000
#> 138  chr2 97211 97218 chr2:0-100000
#> 139  chr2 98156 98163 chr2:0-100000