Find positions of query seq matches in seq column

Find positions of query seq matches in seq column

find_motifs(sequence_df, query_seq)

Arguments

query_seq
sequence to identify in df$seq column, can contain regex

df
data frame containing a header and sequence column

Value

A data.frame with chrom, start, end, and header columns. Start and end indicate start point and end point of query_seq match.

Examples

fa_path <- system.file("extdata", "test.fasta", package = "kentr")
query_seq <- "AATAAA[GTC]"

read_genome_seq(fa_path) %>%
  find_motifs(query_seq)

#> # A tibble: 139 x 4
#>    chrom start   end header       
#>    <chr> <dbl> <int> <chr>        
#>  1 chr1  11544 11551 chr1:0-100000
#>  2 chr1  33977 33984 chr1:0-100000
#>  3 chr1  34608 34615 chr1:0-100000
#>  4 chr1  35489 35496 chr1:0-100000
#>  5 chr1  35497 35504 chr1:0-100000
#>  6 chr1  37324 37331 chr1:0-100000
#>  7 chr1  39909 39916 chr1:0-100000
#>  8 chr1  40753 40760 chr1:0-100000
#>  9 chr1  42815 42822 chr1:0-100000
#> 10 chr1  43116 43123 chr1:0-100000
#> # … with 129 more rows